+++ lib/src/intialize_connection.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
-Future<HttpServer> initializeHttpConnection() async {
-  final server = await HttpServer.bind('localhost', 3000);
+Future<HttpServer> initializeHttpConnection({required int port}) async {
+  final server = await HttpServer.bind('localhost', port, shared: true);
   print('Server listening on localhost:3000');
   return server;
 }